+++ lib/src/eneural_net_extension.dart
@@ -5,6 +5,39 @@ import 'package:collection/collection.dart';
 import 'package:eneural_net/eneural_net.dart';
 
 extension Int32x4Extension on Int32x4 {
+  /// Converts to a [Float32x4].
+  Float32x4 toFloat32x4() =>
+      Float32x4(x.toDouble(), y.toDouble(), z.toDouble(), w.toDouble());
+
+  /// Filter this with [mapper].
+  Int32x4 filter(Int32x4 Function(Int32x4 e) mapper) => mapper(this);
+
+  /// Filter each value with [mapper] and return a [Int32x4].
+  Int32x4 filterValues(int Function(int e) mapper) {
+    return Int32x4(
+      mapper(x),
+      mapper(y),
+      mapper(z),
+      mapper(w),
+    );
+  }
+
+  /// Filter each value with [mapper] and return a [Float32x4].
+  Float32x4 filterToDoubleValues(double Function(int e) mapper) {
+    return Float32x4(
+      mapper(x),
+      mapper(y),
+      mapper(z),
+      mapper(w),
+    );
+  }
+
+  /// Map using [mapper].
+  T map<T>(T Function(Int32x4 e) mapper) => mapper(this);
+
+  /// Returns values as `List<int>`.
+  List<int> toInts() => <int>[x, y, z, w];
+
   Int32x4 operator *(Int32x4 other) => Int32x4(
         x * other.x,
         y * other.y,
@@ -19,8 +52,28 @@ extension Int32x4Extension on Int32x4 {
         w ~/ other.w,
       );
 
+  /// Returns the minimal lane value.
+  int get minInLane {
+    var min = x;
+    if (y < min) min = y;
+    if (z < min) min = z;
+    if (w < min) min = w;
+    return min;
+  }
+
+  /// Returns the maximum lane value.
+  int get maxInLane {
+    var max = x;
+    if (y > max) max = y;
+    if (z > max) max = z;
+    if (w > max) max = w;
+    return max;
+  }
+
+  /// Sum lane.
   int get sumLane => x + y + z + w;
 
+  /// Sum part of the lane, until [size].
   int sumLanePartial(int size) {
     switch (size) {
       case 1:
@@ -36,8 +89,10 @@ extension Int32x4Extension on Int32x4 {
     }
   }
 
+  /// Sum lane squares.
   int get sumSquaresLane => (x * x) + (y * y) + (z * z) + (w * w);
 
+  /// Sum part of the lane squares, until [size].
   int sumSquaresLanePartial(int size) {
     switch (size) {
       case 1:
@@ -53,6 +108,7 @@ extension Int32x4Extension on Int32x4 {
     }
   }
 
+  /// Returns true if equals to [other] values.
   bool equalsValues(Int32x4 other) {
     var diff = this - other;
     return diff.x == 0 && diff.y == 0 && diff.z == 0 && diff.w == 0;
@@ -60,8 +116,64 @@ extension Int32x4Extension on Int32x4 {
 }
 
 extension Float32x4Extension on Float32x4 {
+  /// Converts to a [Int32x4].
+  Int32x4 toInt32x4() => Int32x4(x.toInt(), y.toInt(), z.toInt(), w.toInt());
+
+  /// Perform a `toInt()` in each value and return a [Float32x4].
+  Float32x4 toIntAsFloat32x4() => Float32x4(x.toInt().toDouble(),
+      y.toInt().toDouble(), z.toInt().toDouble(), w.toInt().toDouble());
+
+  /// Filter this with [mapper].
+  Float32x4 filter(Float32x4 Function(Float32x4 e) filter) => filter(this);
+
+  /// Filter each value with [mapper] and return a [Float32x4].
+  Float32x4 filterValues(double Function(double e) mapper) {
+    return Float32x4(
+      mapper(x),
+      mapper(y),
+      mapper(z),
+      mapper(w),
+    );
+  }
+
+  /// Filter each value with [mapper] and return a [Int32x4].
+  Int32x4 filterToIntValues(int Function(double e) mapper) {
+    return Int32x4(
+      mapper(x),
+      mapper(y),
+      mapper(z),
+      mapper(w),
+    );
+  }
+
+  /// Map using [mapper].
+  T map<T>(T Function(Float32x4 e) mapper) => mapper(this);
+
+  /// Returns values as `List<double>`.
+  List<double> toDoubles() => <double>[x, y, z, w];
+
+  /// Returns the minimum lane value.
+  double get minInLane {
+    var min = x;
+    if (y < min) min = y;
+    if (z < min) min = z;
+    if (w < min) min = w;
+    return min;
+  }
+
+  /// Returns the maximum lane value.
+  double get maxInLane {
+    var max = x;
+    if (y > max) max = y;
+    if (z > max) max = z;
+    if (w > max) max = w;
+    return max;
+  }
+
+  /// Sum lane.
   double get sumLane => x + y + z + w;
 
+  /// Sum part of the lane, until [size].
   double sumLanePartial(int size) {
     switch (size) {
       case 1:
@@ -77,8 +189,10 @@ extension Float32x4Extension on Float32x4 {
     }
   }
 
+  /// Sum lane squares.
   double get sumSquaresLane => (x * x) + (y * y) + (z * z) + (w * w);
 
+  /// Sum part of the lane squares, until [size].
   double sumSquaresLanePartial(int size) {
     switch (size) {
       case 1:
@@ -94,6 +208,7 @@ extension Float32x4Extension on Float32x4 {
     }
   }
 
+  /// Returns true if equals to [other] values.
   bool equalsValues(Float32x4 other) {
     var diff = this - other;
     return diff.x == 0.0 && diff.y == 0.0 && diff.z == 0.0 && diff.w == 0.0;